pushing some failed commits

+++ app/src/main/java/com/example/foodhub/admin/fragments/list/AdminVolunteerFragment.kt
@@ -8,6 +8,7 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.appcompat.widget.SearchView
+import androidx.core.content.ContentProviderCompat.requireContext
 import androidx.core.view.MenuHost
 import androidx.core.view.MenuProvider
 import androidx.databinding.DataBindingUtil
@@ -18,8 +19,11 @@ import androidx.lifecycle.lifecycleScope
 import androidx.navigation.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.foodhub.R
+import com.example.foodhub.admin.AdminMainActivity
 import com.example.foodhub.admin.adapters.AdminVoluntaryWorkAdapter
 import com.example.foodhub.admin.viewmodels.VoluntaryWorkViewModel
+import com.example.foodhub.database.CheckConnection
+import com.example.foodhub.database.DBSyncManager
 import com.example.foodhub.database.ImageStorageManager
 import com.example.foodhub.database.tables.VoluntaryWork
 import com.example.foodhub.databinding.FragmentAdminVolunteerBinding
@@ -30,6 +34,9 @@ class AdminVolunteerFragment : Fragment(), MenuProvider, SearchView.OnQueryTextL
     private lateinit var bindingAdminVolunteer: FragmentAdminVolunteerBinding
     private lateinit var voluntaryWorkViewModel: VoluntaryWorkViewModel
     private lateinit var adapter: AdminVoluntaryWorkAdapter
+    private val checkConnection by lazy { CheckConnection((activity as AppCompatActivity).application) }
+
+    private var internetUp = false
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -49,7 +56,23 @@ class AdminVolunteerFragment : Fragment(), MenuProvider, SearchView.OnQueryTextL
         menuHost.addMenuProvider(this, viewLifecycleOwner, Lifecycle.State.RESUMED)
 
         bindingAdminVolunteer.addVolunteer.setOnClickListener { view: View ->
-            view.findNavController().navigate(R.id.adminAddVolunteerFragment)
+
+            checkConnection.observe((activity as AppCompatActivity)) {
+                if (it) {
+                    internetUp = true
+                }
+            }
+
+            if (internetUp) {
+                view.findNavController().navigate(R.id.adminAddVolunteerFragment)
+            }
+            else {
+                val builder = AlertDialog.Builder(requireContext())
+                builder.setPositiveButton("OK") { _, _ -> }
+                builder.setTitle("No Internet Connection")
+                builder.setMessage("Please try again later")
+                builder.create().show()
+            }
         }
 
         adapter = AdminVoluntaryWorkAdapter()
@@ -71,38 +94,73 @@ class AdminVolunteerFragment : Fragment(), MenuProvider, SearchView.OnQueryTextL
 
     override fun onMenuItemSelected(menuItem: MenuItem): Boolean {
         if (menuItem.itemId == R.id.menu_delete_all) {
-            deleteAllWork()
+
+            checkConnection.observe((activity as AppCompatActivity)) {
+                if (it) {
+                    internetUp = true
+                }
+            }
+
+            if (internetUp) {
+                val builder = AlertDialog.Builder(requireContext())
+                builder.setPositiveButton("Yes") { _, _ ->
+                    deleteAllWork()
+                    Toast.makeText(
+                        requireContext(),
+                        "Successfully removed all work entries",
+                        Toast.LENGTH_SHORT).show()
+                }
+                builder.setNegativeButton("No") { _, _ -> }
+                builder.setTitle("Delete all work entries?")
+                builder.setMessage("Are you sure you want to delete all work entries?")
+                builder.create().show()
+            }
+            else {
+                val builder = AlertDialog.Builder(requireContext())
+                builder.setPositiveButton("OK") { _, _ -> }
+                builder.setTitle("No Internet Connection")
+                builder.setMessage("Please try again later")
+                builder.create().show()
+            }
+        }
+        else if (menuItem.itemId == R.id.menu_refresh) {
+
+            checkConnection.observe((activity as AppCompatActivity)) {
+                if (it) {
+                    internetUp = true
+                }
+            }
+
+            if (internetUp) {
+                deleteAllWork()
+                DBSyncManager.syncWorkFromRemote(viewLifecycleOwner, requireContext())
+            }
+            else {
+                val builder = AlertDialog.Builder(requireContext())
+                builder.setPositiveButton("OK") { _, _ -> }
+                builder.setTitle("No Internet Connection")
+                builder.setMessage("Please try again later")
+                builder.create().show()
+            }
         }
         return false
     }
 
     private fun deleteAllWork() {
-        val builder = AlertDialog.Builder(requireContext())
-        builder.setPositiveButton("Yes") { _, _ ->
-
-            // Delete image files from internal storage upon entry removal from ROOM
-            var vWorkList = emptyList<VoluntaryWork>()
-            val imageFileNames: MutableList<String> = mutableListOf()
-            voluntaryWorkViewModel.getAllWork.observe(viewLifecycleOwner, Observer { list ->
-                vWorkList = list
-            })
-            for (item in vWorkList) {
-                imageFileNames.add(item.vImage.substring(item.vImage.lastIndexOf("/")+1))
-            }
-            viewLifecycleOwner.lifecycleScope.launch {
-                ImageStorageManager.deleteAllImagesFromInternalStorage(requireContext(), imageFileNames)
-            }
-
-            voluntaryWorkViewModel.deleteAllWork()
-            Toast.makeText(
-                requireContext(),
-                "Successfully removed all work entries",
-                Toast.LENGTH_SHORT).show()
+        // Delete image files from internal storage upon entry removal from ROOM
+        var vWorkList = emptyList<VoluntaryWork>()
+        val imageFileNames: MutableList<String> = mutableListOf()
+        voluntaryWorkViewModel.getAllWork.observe(viewLifecycleOwner, Observer { list ->
+            vWorkList = list
+        })
+        for (item in vWorkList) {
+            imageFileNames.add(item.vImage.substring(item.vImage.lastIndexOf("/")+1))
         }
-        builder.setNegativeButton("No") { _, _ -> }
-        builder.setTitle("Delete all work entries?")
-        builder.setMessage("Are you sure you want to delete all work entries?")
-        builder.create().show()
+        viewLifecycleOwner.lifecycleScope.launch {
+            ImageStorageManager.deleteAllImagesFromInternalStorage(requireContext(), imageFileNames)
+        }
+
+        voluntaryWorkViewModel.deleteAllWork()
     }
 
     override fun onQueryTextSubmit(query: String?): Boolean {
@@ -133,6 +191,19 @@ class AdminVolunteerFragment : Fragment(), MenuProvider, SearchView.OnQueryTextL
     override fun onResume() {
         super.onResume()
         (activity as AppCompatActivity).findViewById<TextView>(R.id.admin_toolbar_title).text = "Voluntary Works"
-    }
 
+        bindingAdminVolunteer.apply {
+            checkConnection.observe(viewLifecycleOwner) {
+                if (it) {
+                    (activity as AppCompatActivity).findViewById<TextView>(R.id.admin_no_internet_bar).visibility = View.GONE
+                    //internetUp = true
+
+                }
+                else {
+                    (activity as AppCompatActivity).findViewById<TextView>(R.id.admin_no_internet_bar).visibility = View.VISIBLE
+                    //internetUp = false
+                }
+            }
+        }
+    }
 }
